feat(immutable-data): Created new rule.
This rule is a merge of no-object-mutation and no-array-mutation.

re #11
