+++ src/rules/immutable-data.ts
@@ -4,6 +4,11 @@ import { JSONSchema4 } from "json-schema";
 
 import * as ignore from "../common/ignore-options";
 import {
+  AssumeTypesOption,
+  assumeTypesOptionSchema
+} from "../common/types-options";
+import { isExpected } from "../util/misc";
+import {
   checkNode,
   createRule,
   getTypeOfNode,
@@ -11,6 +16,7 @@ import {
   RuleMetaData,
   RuleResult
 } from "../util/rule";
+import { inConstructor } from "../util/tree";
 import {
   isArrayConstructorType,
   isArrayExpression,
@@ -18,17 +24,18 @@ import {
   isCallExpression,
   isIdentifier,
   isMemberExpression,
-  isNewExpression
+  isNewExpression,
+  isObjectConstructorType
 } from "../util/typeguard";
 
 // The name of this rule.
-export const name = "no-array-mutation" as const;
+export const name = "immutable-data" as const;
 
 // The options this rule can take.
 type Options = readonly [
   ignore.IgnorePatternOption &
     ignore.IgnoreAccessorPatternOption &
-    ignore.IgnoreNewArrayOption
+    AssumeTypesOption
 ];
 
 // The schema for the rule options.
@@ -36,27 +43,29 @@ const schema: JSONSchema4 = [
   deepMerge([
     ignore.ignorePatternOptionSchema,
     ignore.ignoreAccessorPatternOptionSchema,
-    ignore.ignoreNewArrayOptionSchema
+    assumeTypesOptionSchema
   ])
 ];
 
 // The default options for the rule.
 const defaultOptions: Options = [
   {
-    ignoreNewArray: true
+    assumeTypes: true
   }
 ];
 
 // The possible error messages.
 const errorMessages = {
-  generic: "Mutating an array is not allowed."
+  generic: "Modifying an existing object/array is not allowed.",
+  object: "Modifying properties of existing object not allowed.",
+  array: "Modifying an array is not allowed."
 } as const;
 
 // The meta data for this rule.
 const meta: RuleMetaData<keyof typeof errorMessages> = {
   type: "suggestion",
   docs: {
-    description: "Disallow mutating arrays.",
+    description: "Enforce treating data as immutable.",
     category: "Best Practices",
     recommended: "error"
   },
@@ -65,11 +74,11 @@ const meta: RuleMetaData<keyof typeof errorMessages> = {
 };
 
 /**
- * Methods that mutate an array.
+ * Array methods that mutate an array.
  *
  * @see https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/Array/prototype#Methods#Mutator_methods
  */
-const mutatorMethods = [
+const arrayMutatorMethods = [
   "copyWithin",
   "fill",
   "pop",
@@ -82,12 +91,12 @@ const mutatorMethods = [
 ] as const;
 
 /**
- * Methods that return a new array without mutating the original.
+ * Array methods that return a new object (or array) without mutating the original.
  *
  * @see https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/Array/prototype#Methods#Accessor_methods
  * @see https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/Array/prototype#Iteration_methods
  */
-const newArrayReturningMethods = [
+const arrayNewObjectReturningMethods = [
   "concat",
   "slice",
   "filter",
@@ -97,14 +106,14 @@ const newArrayReturningMethods = [
 ] as const;
 
 /**
- * Functions that create a new array.
+ * Array constructor functions that create a new array.
  *
  * @see https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/Array#Methods
  */
-const constructorFunctions = ["from", "of"] as const;
+const arrayConstructorFunctions = ["from", "of"] as const;
 
 /**
- * Check if the given node violates this rule.
+ * Check if the given assignment expression violates this rule.
  */
 function checkAssignmentExpression(
   node: TSESTree.AssignmentExpression,
@@ -114,7 +123,8 @@ function checkAssignmentExpression(
     context,
     descriptors:
       isMemberExpression(node.left) &&
-      isArrayType(getTypeOfNode(node.left.object, context))
+      // Ignore if in a constructor - allow for field initialization.
+      !inConstructor(node)
         ? [{ node, messageId: "generic" }]
         : []
   };
@@ -130,9 +140,7 @@ function checkUnaryExpression(
   return {
     context,
     descriptors:
-      node.operator === "delete" &&
-      isMemberExpression(node.argument) &&
-      isArrayType(getTypeOfNode(node.argument.object, context))
+      node.operator === "delete" && isMemberExpression(node.argument)
         ? [{ node, messageId: "generic" }]
         : []
   };
@@ -147,11 +155,9 @@ function checkUpdateExpression(
 ): RuleResult<keyof typeof errorMessages, Options> {
   return {
     context,
-    descriptors:
-      isMemberExpression(node.argument) &&
-      isArrayType(getTypeOfNode(node.argument.object, context))
-        ? [{ node, messageId: "generic" }]
-        : []
+    descriptors: isMemberExpression(node.argument)
+      ? [{ node, messageId: "generic" }]
+      : []
   };
 }
 
@@ -163,30 +169,50 @@ function checkCallExpression(
   context: RuleContext<keyof typeof errorMessages, Options>,
   [options]: Options
 ): RuleResult<keyof typeof errorMessages, Options> {
+  const assumeTypesForArrays =
+    options.assumeTypes === true ||
+    (options.assumeTypes !== false && Boolean(options.assumeTypes.forArrays));
+  const assumeTypesForObjects =
+    options.assumeTypes === true ||
+    (options.assumeTypes !== false && Boolean(options.assumeTypes.forObjects));
+
   return {
     context,
     descriptors:
-      isMemberExpression(node.callee) &&
-      isIdentifier(node.callee.property) &&
-      mutatorMethods.some(
-        m =>
-          m ===
-          ((node.callee as TSESTree.MemberExpression)
-            .property as TSESTree.Identifier).name
-      ) &&
-      (!options.ignoreNewArray ||
-        !isInChainCallAndFollowsNew(node.callee, context)) &&
-      isArrayType(getTypeOfNode(node.callee.object, context))
-        ? [{ node, messageId: "generic" }]
+      // Potential object mutation?
+      isMemberExpression(node.callee) && isIdentifier(node.callee.property)
+        ? // Potential array mutation?
+          arrayMutatorMethods.some(
+            m =>
+              m ===
+              ((node.callee as TSESTree.MemberExpression)
+                .property as TSESTree.Identifier).name
+          ) &&
+          !isInChainCallAndFollowsNew(
+            node.callee,
+            context,
+            assumeTypesForArrays
+          ) &&
+          isArrayType(
+            getTypeOfNode(node.callee.object, context),
+            assumeTypesForArrays,
+            node.callee.object
+          )
+          ? [{ node, messageId: "array" }]
+          : // Potential non-array object mutation (Object.assign on identifier)?
+          node.callee.property.name === "assign" &&
+            node.arguments.length >= 2 &&
+            (isIdentifier(node.arguments[0]) ||
+              isMemberExpression(node.arguments[0])) &&
+            isObjectConstructorType(
+              getTypeOfNode(node.callee.object, context),
+              assumeTypesForObjects,
+              node.callee.object
+            )
+          ? [{ node, messageId: "object" }]
+          : []
         : []
   };
-}
-
-/**
- * Returns a function that checks if the given value is the same as the expected value.
- */
-function isExpected<T>(expected: T): (actual: T) => boolean {
-  return actual => actual === expected;
 }
 
 /**
@@ -198,26 +224,33 @@ function isExpected<T>(expected: T): (actual: T) => boolean {
  */
 function isInChainCallAndFollowsNew(
   node: TSESTree.MemberExpression,
-  context: RuleContext<keyof typeof errorMessages, Options>
+  context: RuleContext<keyof typeof errorMessages, Options>,
+  assumeArrayTypes: boolean
 ): boolean {
   return (
     // Check for: [0, 1, 2]
     isArrayExpression(node.object) ||
     // Check for: new Array()
     ((isNewExpression(node.object) &&
-      isArrayConstructorType(getTypeOfNode(node.object.callee, context))) ||
+      isArrayConstructorType(
+        getTypeOfNode(node.object.callee, context),
+        assumeArrayTypes,
+        node.object.callee
+      )) ||
       (isCallExpression(node.object) &&
         isMemberExpression(node.object.callee) &&
         isIdentifier(node.object.callee.property) &&
         // Check for: Object.from(iterable)
-        ((constructorFunctions.some(
+        ((arrayConstructorFunctions.some(
           isExpected(node.object.callee.property.name)
         ) &&
           isArrayConstructorType(
-            getTypeOfNode(node.object.callee.object, context)
+            getTypeOfNode(node.object.callee.object, context),
+            assumeArrayTypes,
+            node.object.callee.object
           )) ||
           // Check for: array.slice(0)
-          newArrayReturningMethods.some(
+          arrayNewObjectReturningMethods.some(
             isExpected(node.object.callee.property.name)
           ))))
   );
@@ -247,6 +280,8 @@ export const rule = createRule<keyof typeof errorMessages, Options>({
       ignoreOptions,
       otherOptions
     );
+    // This functionality is only avaliable if the parser services are
+    // avaliable.
     const _checkCallExpression = checkNode(
       checkCallExpression,
       context,
