+++ src/rules/index.ts
@@ -1,7 +1,7 @@
 import {
-  name as noArrayMutationRuleName,
-  rule as noArrayMutationRule
-} from "./no-array-mutation";
+  name as immutableDataRuleName,
+  rule as immutableDataRule
+} from "./immutable-data";
 import { name as noClassRuleName, rule as noClassRule } from "./no-class";
 import { name as noDeleteRuleName, rule as noDeleteRule } from "./no-delete";
 import {
@@ -18,14 +18,6 @@ import {
   rule as noLoopRule
 } from "./no-loop-statement";
 import {
-  name as noObjectMutationRuleName,
-  rule as noObjectMutationRule
-} from "./no-object-mutation";
-import { name as noRejectRuleName, rule as noRejectRule } from "./no-reject";
-import { name as noThisRuleName, rule as noThisRule } from "./no-this";
-import { name as noThrowRuleName, rule as noThrowRule } from "./no-throw";
-import { name as noTryRuleName, rule as noTryRule } from "./no-try";
-import {
   name as noMethodSignatureRuleName,
   rule as noMethodSignatureRule
 } from "./no-method-signature";
@@ -33,6 +25,10 @@ import {
   name as noMixedInterfaceRuleName,
   rule as noMixedInterfaceRule
 } from "./no-mixed-interface";
+import { name as noRejectRuleName, rule as noRejectRule } from "./no-reject";
+import { name as noThisRuleName, rule as noThisRule } from "./no-this";
+import { name as noThrowRuleName, rule as noThrowRule } from "./no-throw";
+import { name as noTryRuleName, rule as noTryRule } from "./no-try";
 import {
   name as readonlyArrayRuleName,
   rule as readonlyArrayRule
@@ -46,7 +42,7 @@ import {
  * All of the custom rules.
  */
 export const rules = {
-  [noArrayMutationRuleName]: noArrayMutationRule,
+  [immutableDataRuleName]: immutableDataRule,
   [noClassRuleName]: noClassRule,
   [noDeleteRuleName]: noDeleteRule,
   [noExpressionStatementRuleName]: noExpressionStatementRule,
@@ -55,7 +51,6 @@ export const rules = {
   [noLoopRuleName]: noLoopRule,
   [noMethodSignatureRuleName]: noMethodSignatureRule,
   [noMixedInterfaceRuleName]: noMixedInterfaceRule,
-  [noObjectMutationRuleName]: noObjectMutationRule,
   [noRejectRuleName]: noRejectRule,
   [noThisRuleName]: noThisRule,
   [noThrowRuleName]: noThrowRule,
