+++ src/util/typeguard.ts
@@ -155,27 +155,93 @@ export function isUnionType(type: Type): type is UnionType {
   return ts !== undefined && type.flags === ts.TypeFlags.Union;
 }
 
-export function isArrayType(type: Type): type is ArrayType {
-  return (
-    (type.symbol && type.symbol.name === "Array") ||
-    (isUnionType(type) && type.types.some(isArrayType))
-  );
+export function isArrayType(type: Type | null): type is ArrayType;
+export function isArrayType(
+  type: Type,
+  assumeType: false,
+  node: null
+): type is ArrayType;
+export function isArrayType(
+  type: Type | null,
+  assumeType: boolean,
+  node: TSESTree.Node | null
+): type is ArrayType;
+export function isArrayType(
+  type: null,
+  assumeType: true,
+  node: TSESTree.Node
+): boolean;
+export function isArrayType(
+  type: Type | null,
+  assumeType: boolean = false,
+  node: TSESTree.Node | null = null
+): boolean {
+  return assumeType === true
+    ? node !== null
+    : type !== null &&
+        ((type.symbol && type.symbol.name === "Array") ||
+          (isUnionType(type) &&
+            type.types.some(t => isArrayType(t, false, null))));
 }
 
 export function isArrayConstructorType(
-  type: Type
-): type is ArrayConstructorType {
-  return (
-    (type.symbol && type.symbol.name === "ArrayConstructor") ||
-    (isUnionType(type) && type.types.some(isArrayConstructorType))
-  );
+  type: Type | null
+): type is ArrayConstructorType;
+export function isArrayConstructorType(
+  type: Type,
+  assumeType: false,
+  node: null
+): type is ArrayConstructorType;
+export function isArrayConstructorType(
+  type: Type | null,
+  assumeType: boolean,
+  node: TSESTree.Node | null
+): type is ArrayConstructorType;
+export function isArrayConstructorType(
+  type: null,
+  assumeType: true,
+  node: TSESTree.Node
+): boolean;
+export function isArrayConstructorType(
+  type: Type | null,
+  assumeType: boolean = false,
+  node: TSESTree.Node | null = null
+): boolean {
+  return assumeType === true
+    ? node !== null && isIdentifier(node) && node.name === "Array"
+    : type !== null &&
+        ((type.symbol && type.symbol.name === "ArrayConstructor") ||
+          (isUnionType(type) &&
+            type.types.some(t => isArrayConstructorType(t, false, null))));
 }
 
 export function isObjectConstructorType(
-  type: Type
-): type is ObjectConstructorType {
-  return (
-    (type.symbol && type.symbol.name === "ObjectConstructor") ||
-    (isUnionType(type) && type.types.some(isObjectConstructorType))
-  );
+  type: Type | null
+): type is ObjectConstructorType;
+export function isObjectConstructorType(
+  type: Type,
+  assumeType: false,
+  node: null
+): type is ObjectConstructorType;
+export function isObjectConstructorType(
+  type: Type | null,
+  assumeType: boolean,
+  node: TSESTree.Node | null
+): type is ObjectConstructorType;
+export function isObjectConstructorType(
+  type: null,
+  assumeType: true,
+  node: TSESTree.Node
+): boolean;
+export function isObjectConstructorType(
+  type: Type | null,
+  assumeType: boolean = false,
+  node: TSESTree.Node | null = null
+): boolean {
+  return assumeType === true
+    ? node !== null && isIdentifier(node) && node.name === "Object"
+    : type !== null &&
+        ((type.symbol && type.symbol.name === "ObjectConstructor") ||
+          (isUnionType(type) &&
+            type.types.some(t => isObjectConstructorType(t, false, null))));
 }
